requirements.txt
Django==3.2.13
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
importlib-metadata==4.8.3
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-decouple==3.3
pytz==2022.1
requests==2.27.1
urllib3==1.26.9
